-move compat mode to base -implement for pigpio
